Add self-scheduling to threaded translation (vs static)
For the x86-64 sandboxed translator it makes about a 5%
difference over a set of 13 pexes from the web / webstore.
For python.pexe this reduces the translation time w/ 4
threads from 43 seconds to 37 seconds. Most of the other
web benchmarks only get slight improvements (if any).

For the ARM sandboxed translator, it's about 4.5%
aggregate difference for on the a15. Again, most of the
apps are a wash except for python, gamecake and the
gameboy emulator.

For the unsandboxed translator, this is often a wash,
except on a few benchmarks.

Testing on spec2k: Same or better for sandboxed translator.
Notably, crafty translation time drops from 1.59 secs to
1.19 secs w/ CPU utilization going from 1.8 to 2.7
w/ 4 threads. With two threads it's 2.1 secs to 1.6 secs
and CPU utilization from 1.2 to 1.7. Crafty was one of
the benchmarks that regressed after switching from 3.3
to 3.4 and perhaps part of it was bad scheduling.
The gcc and eon benchmarks also get a nice boost.
Similar on ARM: crafty drops from 5.7 to 4.8 w/ 2 threads.

Use a ChunkSize of 8, to reduce the synchronization
overhead somewhat, hopefully without hurting streaming
translation. This tapers off to 1 when there are
few functions remaining to attempt finer grained load
balancing near the end.

For some idea of how this affects streaming translation
the average function size for the benchmarks range
from 160 bytes to 1700 bytes. So 8 times that is
about 1.3KB to 14KB. Geomean of average func size is
570 bytes. For that geomean, 4 threads times 8 functions
would grab ~18KB at a time on geo-average-ish.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3777
R=jfb@chromium.org, stichnot@chromium.org

Review URL: https://codereview.chromium.org/196793026