PNaCl: Add option to turn off use of __aeabi_* functions for Non-SFI NaCl
Currently, PNaCl's arm-nonsfi target uses the triple
"armv7a-linux-none", which uses soft-float, which isn't the right ABI
for Non-SFI NaCl.

Switching to a triple that uses hard-float
(e.g. "armv7a-linux-gnueabihf") has the side effect of causing llc to
generate uses of __aeabi_* functions, which PNaCl's libgcc.a (from
compiler-rt) doesn't provide.

In the longer term, the cleaner way to address that is to build
compiler-rt to provide __aeabi_* functions, but doing that for SFI NaCl
would involve porting some ARM assembly code to SFI NaCl, and that might
be a minor performance regression -- e.g. lib/arm/aeabi_uidivmod.S
defines __aeabi_uidivmod() as a wrapper for __udivmodsi4().

For now, just extend the existing localmod so that we can use it for
Non-SFI NaCl too.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3809
TEST=also tested with pnacl-translate.py change
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/266913009