Generalize pnacl-bccompress to be able to remove existing abbreviations.
Adds a flag to pnacl-bccompress. When true, the input file is copied but
all abbreviations are removed. This allows us to generate simple
bitcode files (for examples in the PNaCl bitcode file ABI), as well
as determine how effective existing abbreviations are in pnacl-freeze.

BUG=None
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/290833002