Add support for building subzero with LLVM
Since subzero depends on LLVM components, it makes sense to build it
along with the LLVM build. Also, toolchain_build doesn't have support
for a target to depend on another target's working directory. We may want
to fix that in the future, but for now it's simpler to do the build
together with LLVM.

This CL allows subzero to be built from an external directory similar to
the way clang is. However we just use a Make variable directly insetad
of a configure flag to avoid regenerating autoconf files.

R=jvoung@chromium.org, stichnot@chromium.org
BUG=

Review URL: https://codereview.chromium.org/299753004