Add PassManager to pnacl-abicheck and pnacl-llc
I need to get a DataLayout in the translator (to deal with vector alignment),
and this requires a PassManager. There are other unexpected dependencies that
crop up when trying to go halfway without adding a PassManager (e.g. the
AsmPrinter is null and crashes on the mangler), but in the end the following
line is the one that requires the PassManager:
  include/llvm/PassAnalysisSupport.h:200

R=jvoung@chromium.org, dschuff@chromium.org
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3870
NOTRY=true

Review URL: https://codereview.chromium.org/326513002