PNaCl sandboxed translator: fix crash
I added a data layout usage to the translator [*] and broke the sandboxed
translator because the function ABI verification pass was added before the pass
manager contained a data layout, causing a default data layout to be constructed
(which errors out). This CL fixes the breakage by making it possible to use data
layout in the translator, even when sandboxed.

  [*]: https://codereview.chromium.org/321733002/

This CL also stubs out LLVM's current_path function because getcwd returns an
error when run inside the sandboxed translator (it does go through a NaCl
syscall, but we don't override it to anything meaningful for the
translator). This happens only in debug builds (which I was testing with) because
it's only asserted on (so release builds don't see this error).

R=jvoung@chromium.org, dschuff@chromium.org
TEST= ./pnacl/build.sh translator-all && ./scons bitcode=1 use_sandboxed_translator=1 small_tests
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3870

Review URL: https://codereview.chromium.org/337793002