PNaCl: Make PNaClABIVerifyModule complain if _start() isn't defined
Make the ABI checker stricter so that it complains if the module doesn't
contain an entry point.

If we mistakenly internalised _start() (but left its name in place),
previously we might not have caught that because pnacl-llc
re-externalises symbols when doing streaming translation.  This new
check would catch the mistake.

Update some tests to define _start() so that they still pass.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3873
TEST=lit tests
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/330623003