Remove special nacltrap handling from X86 NaCl.
I believe it was added previously to work around the validator not
accepting ud2, so it would emit a null pointer deref instead.

However the validator is fixed now:
https://code.google.com/p/nativeclient/issues/detail?id=2094

There was some asm parsing support for parsing a "nacltrap" pseudo-inst,
but the -filetype=asm path never emitted a "nacltrap" since the nacl-as
didn't support that pseudo-inst. I.e., it was supported on the parsing
side but not the writing side. It doesn't look like anything used this
parsing support, and it wouldn't have worked with nacl-as anyway.

BUG=none
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/343273003