Make x86-64 sandbox-base-hiding call sequence work with PIC
The call sequence used on x86-64 to avoid storing the sandbox base
address on the stack was not PIC-friendly because it pushed the return
address as an immediate. When generating PIC code, manually calculate
the return address into a register and push that.

This includes a tiny cherry-pick of upstream r202418, which adds an
accessor for the relocation model to MCObjectFileInfo

R=mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3873
TEST= LLVM regression

Review URL: https://codereview.chromium.org/357603006