Reduce redundancy created by PNaCl's vector globalization pass
This change makes it easier to reconstruct shuffles, and reduces the pass'
reliance on constant merging. It's a toolchain only fix, meaning that no
translator update is needed, only a recompile.

The loads of constant vectors are still present and will be handled in a
separate change.

R=dschuff@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3893
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)

Review URL: https://codereview.chromium.org/379013002