Translator: rematerialize constant vector loads
PNaCl's ABI doesn't contain constant vectors, they therefore need to be
rematerialized by the translator if we want to get the best optimizations
possible.

The CL also renames CombineVectorInstructions.cpp to BackendCanonicalize.cpp to
be more accurate about what the pass does: it can now affect more than vectors
because the above change affects bitcasts and loads in general (since it doesn't
make sense to only stick with vectors). Things should stay in this one pass so
that the translator can restrict itself to what's important to optimize, and do
so as fast as it can.

R=dschuff@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3893
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)

Review URL: https://codereview.chromium.org/385073012