Speed up construction of global initialzers in the PNaCl bitcode reader.
Discovered that the replacement stategy for placeholder forward address
references for global variables, when parsing the bitcode, can be
very non-linear. This is due to the fact that LLVM IR creates unique
constants (i.e. no duplicates). As a result, the same initializer
may be copied several times, as each instance of a placeholder is
replaced.

This change guarantees that we only generate one copy of each
initializer, no matter how many placeholders appear in the
initializer.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3908
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/425113002