Fix an issue which allowed memory access outside a MinSFI sandbox.
Applying the ExpandGEP optimization on memory intrinsics would allow to
set the src/dest addresses at the end of the guard region and access
the following 4GB of memory. The optimization is now applied only on
instructions manipulating first class values and the offset is limited
so that access beyond the guard region is not possible.

BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/432913003