Add support for MinSFI address subspaces smaller than 32 bits.
The prior implementation of MinSFI sandboxing assumed that the sandbox
will have access to 4GB of memory and hence cast pointers to i32.
This patch adds bit masking in order to further limit the size of
the addressable memory. This is necessary for supporting 32-bit
architectures.

BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)
      manually tested on zlib compiled with PNaCl

R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/451523002