Tweak GlobalizeConstantVectors so it doesn't replace ops of consts with non-consts.
If a constant's op is replaced by an instruction, it's
not really a constant anymore. See: https://codereview.chromium.org/324853003/
for a problem this could tickle further down the line.

A simpler example of a problem, is that the pass currently
attempts to replace constant exprs that are nested within
global initializers, with the materializing load
instruction's value. The pass only filters out
GlobalVariable users, so misses the nested case.

So, change the check for GlobalVariables into a broader
check for Constants.

Also  GlobalizeConstantVectors currently only inspects
instruction operands at the top level. It doesn't find
nested constant exprs with vector constants. So,
run -expand-constant-expr earlier, to reduce the nesting
and help GlobalizeConstantVectors find additional
constants.

R=jfb@chromium.org
TBR=jfb@chromium.org
(Lgtm through email, but didn't seem to register)

BUG= https://code.google.com/p/nativeclient/issues/detail?id=2205

tangentially for:
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3609

Review URL: https://codereview.chromium.org/516543002