Initialize untrusted stack pointer in MinSFI's entry function
The runtime of MinSFI copies arguments of the sandbox to the bottom of
the untrusted stack and invokes the sandbox with the corresponding
pointer. This patch changes the ExpandAllocas pass so that it initializes
the untrusted stack pointer at the beginning of the entry function,
setting its value to the pointer argument which happens to be the top of
the stack at that time.

BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/542873002