Report invalid function references in MinSFI's SandboxIndirectCalls pass
Calling a function with the wrong signature by casting its pointer is
illegal in MinSFI's control flow integrity enforcement and will make the
sandoxed code crash by design. However, a bitcast of an inherent function
pointer (a reference directly to the global value) would also crash the
LLVM pass itself. This patch fixes the bug and the pass now reports such
bitcasts. Note that the pass will not identify call sites which cast the
i32 pointer to a wrong function type.

The patch also reports an error if a function is used as a call argument
instead of just asserting it.

BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= (cd ./toolchain_build/out/llvm_i686_linux_work/ && ninja check)
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/553723002