Fix PIC indirect call sequence with sandbox-base hiding
When making indirect calls with sandbox-base hiding, the call sequence
pushes the return address on the stack (using r10) and jumps to the
call target through r11. This sequence clobbers r10, which is OK because
r10 is call-clobbered. However if the call target is stored in r10
before the call, it must be copied into r11 before the indirect jump
sequence.
This CL puts the mov of the call target into r11 before the return-
address push.

R=mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3958

Review URL: https://codereview.chromium.org/620733002