Remove nacltlsaddr32 from assembly layer
nacltlsaddr32 was a pseudo-instruction that essentially duplicated the
existing tlsaddr32 but was parsed by the assembler and expanded in the
MC layer rather than being expanded in X86MCInstLower.cpp. This was to ensure
that the TLS GD sequence stayed bundle-locked. Now that both assemblers
support .bundle_lock we can emit that in X86MCInstLower directly and
remove the special MC support.

However we can't turn that on everywhere yet because gas has a bug when
a call instruction is used in an explicit bundle-locked group,
so we work around it for now. See the comment in
LowerTlsAddr for more details.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3966
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4008

Review URL: https://codereview.chromium.org/764813005