Use "." instead of label difference for GOT address calculations with gas
LLVM's MC layer has no way to generate a "." symbol in assembly files
(or its equivalent for direct object emission). When calculating the GOT's
address, it emits a label and refers to it, but this doesn't work for NaCl
if the instruction after the label gets nop-padded. For this case, we
directly emit the assembly text with "." instead of using the MC layers
(but only when emitting text assembly).

R=mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3982

Review URL: https://codereview.chromium.org/774433002