Fix error reporting in the PNaCl bitcode reader.
When LLVM 3.5 was merged, the handling of errors was not fixed.
The effect is that error messages do not percolate up to the
top-level routines (Which expect: ErrorOr<Module*>). Rather,
pnacl-llc (and similar reading tools) core dump if there is
an error in the bitcode file.

This CL fixes this issue.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4006
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/770853002