Fix a couple of warnings about unused private fields, use TRI, etc.
Also, while I noticed that the TRI for NaClRewritePass is
unused, it looks like that can be used to check if a
register is modified (or its super register is modified),
so use that instead of separately checking ESP & RSP mods.

BUG=none

Just happened to notice while searching for errors in
a broken merge build.

R=dschuff@chromium.org, kschimpf@google.com

Review URL: https://codereview.chromium.org/775393002