Remove localmod for disabling libcall simplifications while compiling implementations of libs.
If I remember correctly, the original problem was that
exp2() is implemented as a call to pow(). However if
the optimizer does XYZ to pow's implementation... then
there is trouble?

A couple of recommendations were to use some -fno-builtin-*
flag or some other way of not relying on TLI when compiling
library code:
http://lists.cs.uiuc.edu/pipermail/llvm-commits/Week-of-Mon-20121210/158679.html
Newlib is built with -fno-builtin.

An earlier problem with, -fno-builtin was that it doesn't
help with LTO because the commandline flags are lost by
then. If the commandline options are embedded in the
bitcode it might help. However, we also explicitly disable
simplify libcalls during LTO in the driver, because
static linking plus simplify libcalls can cause other
problems. E.g., It may mean that some library code is not
linked in, so transforming calls from one lib func to
another after linking can break.

BUG= none (I ran into a merge conflict with this, where
code is inlined, so wanted to see if it can just be
removed)

R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/753533004