Cherry-pick: Use 32-bit ebp for NaCl64 in a limited case: llvm.frameaddress.
When merging up further, will need to fix up the way the
RegisterInfo is accessed. Also need to merge in and fix up
the test for the x32 case, since that depends on a
different commit (that was dropped in this cherrypick).

This also differs from the upstream test in that load
sandboxing and base address hiding is actually being done.

Summary:
Follow up to [x32] "Use ebp/esp as frame and stack pointer":
http://reviews.llvm.org/D4617

In that earlier patch, NaCl64 was made to always use rbp.
That's needed for most cases because rbp should hold a full
64-bit address within the NaCl sandbox so that load/stores
off of rbp don't require sandbox adjustment (zeroing the top
32-bits, then filling those by adding r15).

However, llvm.frameaddress returns a pointer and pointers
are 32-bit for NaCl64. In this case, use ebp instead, which
will make the register copy type check. A similar mechanism
may be needed for llvm.eh.return, but is not added in this change.

Test Plan: test/CodeGen/X86/frameaddr.ll

Reviewers: dschuff, nadav

Subscribers: jfb, llvm-commits

Differential Revision: http://reviews.llvm.org/D6514

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@223510 91177308-0d34-0410-b5e6-96231b3b80d8

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4005
R=dschuff@chromium.org, jfb@chromium.org

Review URL: https://codereview.chromium.org/808543002