Merge-upstream: Fix stack frame adjustments to special case NaCl when not LP64.
NaCl64 was always considered ILP32, but it was not considered "not LP64".

After the merge, NaCl64 is considered "not LP64", but some
parts that check for "not LP64" need to be special cased.
For example, there are some upstream special cases for the
StackPtr and FramePtr (the StackPtr field is set to RSP for
NaCl, but ESP for other "not LP64" 64-bit targets like
x32). Change another instance use of RSP/ESP to match the
special cases of "not LP64" || isNaCl64. This is done to
fix an assert later on (in the NaCl rewrite pass).

Add a test that touches naclssp, and naclspadj, which would
have triggered the asserts.

TODO upstream this too, but the test cannot be upstreamed
yet until the stack/frame ptr sandboxing is upstreamed.

BUG=none
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/811203003