Fix some build warnings.
There were warnings for:
  - Using a default label when the switch already handles all enum values cases (LLVM will instead warn when one enum value isn't handled by a switch);
  - Missing virtual dtor when the class has virtual methods;
  - Missing virtual dtor in the base class;
  - A virtual method hiding a parent's overloaded virtual function;
  - Comparing integers of different types (passing constants to gtest).
I also moved some of the virtual to override when that made sense.

I think most of these appeared when we moved to LLVM 3.5 but went unnoticed.

There are still two warnings left in X86ISelLowering.cpp (RegInfo is an unused variable in two places) which I've left as-is because it may hinder jvoung's upcoming merge.

R=kschimpf@google.com, kschimpf@chromium.org
TEST= ninja check
BUG= build warnings

Review URL: https://codereview.chromium.org/836473005