Delete operator= and copy ctor from NaClBitstreamCursor.
It doesn't appear to be used anymore.

A later merge commit will remove this from the non-NaCl
bitstream reader too (Ben Kramer's r217778),
though it's not called out explicitly in the commit message.
http://lists.cs.uiuc.edu/pipermail/llvm-commits/Week-of-Mon-20140915/235335.html

Also, the operator= has to manage all the fields, and a later
commit will be adding fields in the non-NaCl bitstream cursor
to compensate for changes in lib/Support. Having it around will
be more pain to keep in sync:
http://llvm.org/klaus/llvm/commit/ea3c2111f43cd8f67dd44e03cd58d019efa5f634/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4026
TEST= it compiles
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/860563002