Emit atomic memory order other than seq_cst on demand only
The following CL:
  https://codereview.chromium.org/791053006/
Taught PNaCl to support atomic memory orderings other than seq_cst (which was
the only supported memory ordering at launch). This is desirable for
performance, but it causes issues for users when rolling out the new SDK: old
versions of the translator don't accept anything but seq_cst and the canary SDK
emits them, which makes testing harder. This CL makes the SDK default to
emitting only seq_cst memory ordering, but leaves the translator as is (i.e. it
still accepts the new memory ordering). This makes it easier to roll out this
new feature, and users can still opt-in to emitting the new memory order by
specifying -pnacl-memory-order-seq-cst-only=false. The flag currently defaults
to seq_cst only, and for Chrome 43 it will be made to default to false
(i.e. acq/rel/acq_rel/seq_cst can be emitted).

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4029
TEST= make check

Review URL: https://codereview.chromium.org/857403002