Simplify implementation of LLVM bitcode abbreviations.
The previous implementation of bitcode abbreviations treated literals
as a special case. As such, most code had to test if it is a literal, and
apply a switch on all other encodings separately. This change makes
literals just another encoding, letting a single switch statement to describe
actions to be applied to a bitcode abbreviation.

Also adds error checking to bitstream reader when reading
bitcode abbreviations, so that it generates better error
messages if malformed abbreviation operators are found.

BUG=None
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/848983002