Switch NACL_SANDBOX to PNACL_BROWSER_TRANSLATOR, use it consistently.
Replace the build variable NO_SANDBOX with a more clear name:
PNACL_BROWSER_TRANSLATOR. Add a define to allow it to be used
in place of defined(__native_client__) or defined(__pnacl__) when appropriate
inside C++ source.

This allows the subtle distinction between things that are enabled/disabled
because of NaCl's general limitations vs things that are the result of running
inside the PNaCl translator.

This is an incremental step towards building the pnacl tools with
the pnacl tools.

BUG=https://code.google.com/p/naclports/issues/detail?id=130
TEST=None
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/928373005