ExpandStructRegs: ignore intrinsics
My previous patch moved ExpandStructRegs to post-opt only, which caused issues
in the subzero build because we're still missing a pass which fixes struct
registers as function parameters. Until we have this fix we need to run
ExpandStructRegs as a pre-opt pass, and we need to ignore intrinsics which
return struct regs.

R=dschuff@chromium.org, jvoung@chromium.org, mtrofin@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3857

Review URL: https://codereview.chromium.org/994803002