PromoteIntegers: handle constants of illegal size
When self-compiling clang we witness the following IR:
  store i248 1987729907822349521421004355350957462305601531112707179148245392796156160, i248* undef, align 4

Which becomes:
  store i64 1099511628032, i64* undef, align 4
  store i64 4503599627371008, i64* undef, align 4
  store i64 256, i64* undef, align 4
  store i56 316663643766784, i56* undef, align 4

The i56 store was problematic for PromoteIntegers's splitStore because it
expected an instruction, not a constant.

This issue is specific to stores, since in other cases the constants are
handled and size doesn't matter, it just gets adjusted for the instruction.

R=dschuff@chromium.org
BUG= none
TEST= make check

Review URL: https://codereview.chromium.org/996513002