ExpandArithWithOverflow: fix crash
The code was using a dangling Twine, replace it with a StringRef instead. This
issue showed up in the Windows build only because it uses Release without
asserts and GCC. The DEBUG macro was extending the lifetime of the Twine (within
Call). DEBUG does nothing without asserts, which led to further optimization.

R=dschuff@chromium.org
BUG= crash on Windows when doing a toolchain roll

Review URL: https://codereview.chromium.org/992323002