Nativized variables shouldn't consume stack space.
Prior to this change, whether or not a variable had been "nativized"
wasn't taken into account when calculating the static space required
for the frame.

Now, they aren't included in that and a minor safety measure is taken
to ensure that the frame offset of a nativized variable isn't requested.

Additionally, the logic for skipping a stack restoration no longer
emits a stack restore when only nativized variables are present within
the frame.