Fix NaCl's default to .init_array for constructors
We want NaCl to default to .init_array instead of .ctors for constructors.
So we had custom code to set the StaticCtorSection in
TargetLoweringObjectFileNaCl. Meanwhile the code around the LOCALMODs changed
in LLVM 3.6 and broke our setting of the default. This CL updates the LOCALMOD
to set it the way Linux does, and also adds a test.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4120

Review URL: https://codereview.chromium.org/993163003