ExpandVarArgs: don't pack the vararg struct
VarArgs are transformed into alloca of a struct passed onto the stack. Making
that struct packed means that we generate unaligned loads and stores of 64-bit
floating-point values which is suboptimal performance-wise.

This CL changes our approach to rely on the target's (e.g. le32-unknown-nacl
or asmjs-unknown-emscripten) preferred alignment for types within the struct,
and the target's default stack alignment. For NaCl this means 64-bit values
will be 64-bit aligned in a struct, and the stack will be 16-byte aligned, so
the stores and loads will also be aligned.

This was prompted by an Emscripten change, see:
  https://github.com/kripken/emscripten-fastcomp/commit/fa1c90b56aa3697fb0868ccea0befc0be3f89921

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
R=dschuff@chromium.org, sunfish@mozilla.com
TEST= make check

Review URL: https://codereview.chromium.org/1004913005