ExpandVarArgs: refactor to C++11 and use IRBuilder
My previous CL[1] was wrong because it removes the packed attribute from the
struct that replaces the va_list (to avoid unaligned load/store), but doesn't
adjust the loads on the callee side to be aligned as they would be in the
caller's struct packing.

This patch moves the code to C++11 and uses IRBuilder as preparation for fixing
the breakage (which requires ConstantExpr).

  [1]: https://codereview.chromium.org/1004913005/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
R=dschuff@chromium.org, sunfish@mozilla.com
TEST= make check

Review URL: https://codereview.chromium.org/1005383002