ExpandVarArgs: properly align loads from va_arg
A previous CL[1] was wrong because it removes the packed attribute from the
struct that replaces the va_list (to avoid unaligned load/store), but doesn't
adjust the loads on the callee side to be aligned as they would be in the
caller's struct packing. This patch fixes the callee side, and relies on a
refactoring CL[2].

This failure was only caught by the NaCl-side barebones vaarg test which called
a vaarg function with `long` as well as `long long` and read bad data from the
stack. This CL adds a LIT test in expand-varargs.ll to prevent further
regression.

  [1]: https://codereview.chromium.org/1004913005/
  [2]: https://codereview.chromium.org/1005383002/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
R=dschuff@chromium.org, sunfish@mozilla.com
TEST= make check
TEST= ./scons bitcode=1 run_barebones_vaarg_test

Review URL: https://codereview.chromium.org/1015533003