Revert cygwin-related LOCALMOD that disables mmap in MemoryBuffer/file tools.
This reverts a70a76d7dd5:
https://codereview.chromium.org/19923003

The localmod was added in the 3.3 merge, but we don't use
cygwin for the windows LLVM build anymore so the ifdef
doesn't trigger in practice. I guess it may be a problem
if we ever go *back* to cygwin... but...

Happened to notice this while trying to reading through
what MemoryBuffer vs MemoryBufferRef does.

BUG=none
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1014043005