Clean up use of MemoryBuffer in NaClObjDump and NaClBitcodeMunge.
Changes the API to function NaClObjDump to expect a MemoryBufferRef
instead of a MemoryBuffer*, to be consistent with other functions
in NaClReaderWriter.h

Also cleans up class NaClBitcodeMunge.cpp to reuse the low-level
implementation of the memory buffer. This was done after discovering
that getMemBufferCopy allocates (and leaks) an internal low-level
implementation of the memory buffer. The code now reuses a single
buffer over all tests for that class, stopping the leak to memory.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4103
R=dschuff@chromium.org, jvoung@chromium.org

Review URL: https://codereview.chromium.org/1013823002