Removes unused abbreviations from compressed code.
The previous version of pnacl-bccompress never removed abbreviations
as new abbreviations were added. This resulted in many unused
abbreviations being left in the compressed file. This has two costs.
The first is that the unnecessary abbreviation is saved to the
file. The other is that it may cause the bitcode to use more bits for
each abbreviation, increasing the size of each record.

This CL removes such abbreviations by adding another bitcode scan to
find the best abbreviation to use for each record, and then remove
unused abbreviations before writing out the compressed bitcode file.

This change (for large bitcode files) results in
pnacl-compress running about a 10% slower and
produces a bitcode file that is about 1.2% smaller.

BUG=None
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1003303002