Avoid generating register-register load on ARM
LLVM was generating register-register loads on ARM (rejected by the sandbox)
when doing bswap16 because the pattern that combines the load wasn't predicated
on NaCl. This patch adds this predication so that NaCl doesn't generate that
pattern.

The LLVM pattern was added in the following patch:
  https://llvm.org/svn/llvm-project/llvm/trunk@208620

BUG= https://code.google.com/p/chromium/issues/detail?id=460432
R=jvoung@chromium.org
TEST= make check

Review URL: https://codereview.chromium.org/1008233005