ResolvePNaClIntrinsics: properly resolve is_lock_free
Handle non-browser translator case using the Triple instead of the host
architecture, which allows proper cross-compilation. Also add support for the
asmjs target (which currently doesn't really support atomics anyways, but needs
some value and 4 seems likely).

It may be good to query the target at some point in time, but we'd need to get
that information from the embedder, and we don't have it at the moment so it's
still better to hardcode the known platform minimum.

R=dschuff@chromium.org, azakai@mozilla.com
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
TEST= make check
TEST= ./scons bitcode=1 small_tests medium_tests large_tests

Review URL: https://codereview.chromium.org/1047113004